refactor: move components to own files
