+++ src/renderer/shared/ui-entities/ConsensusAccountsList/types.ts
@@ -1,0 +1,4 @@
+import { type Chain } from '@/shared/core';
+import { type AccountId } from '@/shared/polkadotjs-schemas';
+
+export type ChainAccountPair = readonly [chain: Chain, accountId: AccountId];
